Better search.
The input keywords, combined as a single tag, are used to match against tags for each bookmark entry. Partial matches (but in same sequence) are accepted.
The same input keywords are also used to search title metadata. Even if a single keyword is found in URL or title metadata the bookmark is shown in results.

Signed-off-by: Arun Prakash Jana <engineerarun@gmail.com>