Try with http if scheme is missing in URI.
urllib3 requests can work around this, e.g., google.com is resolved.
However, the browser fails if there's no scheme in the URI.
The observation is based on Google Chrome's behaviour.
So we try http://google.com.

Currently there is no way to override this (i.e. try the URI as it is). We'll
need to work on it if there a bug report on this 'forced' behaviour.