Support http proxy environment variables
If defined, the variables http_proxy, HTTP_PROXY, https_proxy or
HTTPS_PROXY will be used to create a proxied pool manager.

If the prefix https:// is used by the url, it will try to use
the corresponding proxy variables. If not defined, it will
fallback to http_proxy and HTTP_PROXY.

Signed-off-by: Denis A. Altoé Falqueto <denisfalqueto@gmail.com>